# src/swagger/Swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pedidos API Digital",
    "description": "API para gestão de pedidos e upload de arquivos TXT desnormalizados.\nAutenticação via Bearer JWT. Datas no formato ISO 8601 (UTC).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Ambiente local"
    },
    {
      "url": "Link não Disponível - Versão Beta",
      "description": "Produção"
    }
  ],
  "security": [{ "BearerAuth": [] }],
  "tags": [
    { "name": "Users", "description": "Registro e gerenciamento de usuários" },
    { "name": "Pedidos", "description": "CRUD e consultas de pedidos" },
    { "name": "Uploads", "description": "Upload e processamento de arquivos TXT" }
  ],
  "paths": {
    "/user/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Registrar um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "username": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Erro ao registrar o usuário" }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "token": { "type": "string" } }
                }
              }
            }
          },
          "400": { "description": "Erro ao realizar login" }
        }
      }
    },
    "/user/lista": {
      "get": {
        "tags": ["Users"],
        "summary": "Obter a lista de usuários",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "username": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Erro ao obter a lista de usuários" }
        }
      }
    },
    "/pedido/lista": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Listar todos os pedidos (sem parâmetros)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Pedido" }
                }
              }
            }
          },
          "500": { "description": "Erro interno" }
        }
      }
    },
    "/pedidos/periodo": {
        "get": {
          "tags": ["Pedidos"],
          "summary": "Listar pedidos por período",
          "security": [{ "BearerAuth": [] }],
          "parameters": [
            {
              "name": "dataInicio",
              "in": "query",
              "required": true,
              "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
              "example": "2025-08-01",
              "description": "Data inicial (YYYY-MM-DD). Também aceita ISO completo."
            },
            {
              "name": "dataFim",
              "in": "query",
              "required": true,
              "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
              "example": "2025-08-19",
              "description": "Data final (YYYY-MM-DD). Também aceita ISO completo."
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de pedidos no período",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": { "type": "boolean" },
                      "message": { "type": "string" },
                      "start": { "type": "string", "format": "date-time" },
                      "end": { "type": "string", "format": "date-time" },
                      "data": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Pedido" }
                      }
                    }
                  }
                }
              }
            },
            "400": { "description": "Query inválida" }
          }
        }
      },
    "/pedido/{id}": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Obter pedido pelo ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pedido" } } }
          },
          "404": { "description": "Pedido não encontrado" }
        }
      },
      "put": {
        "tags": ["Pedidos"],
        "summary": "Atualizar pedido",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PedidoUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido atualizado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pedido" } } }
          },
          "404": { "description": "Pedido não encontrado" }
        }
      },
      "delete": {
        "tags": ["Pedidos"],
        "summary": "Remover pedido",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Pedido removido com sucesso" },
          "404": { "description": "Pedido não encontrado" }
        }
      }
    },
    "/uploads/txt": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Upload de arquivo TXT desnormalizado",        
        "security": [{ "BearerAuth": [] }],        
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Arquivo processado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": { "description": "Requisição inválida" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão" },
          "404": { "description": "Usuário não encontrado" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use o token JWT: 'Bearer {seu_token_aqui}'"
      }
    },
    "parameters": {
      "page": { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
      "limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
      "sort": { "name": "sort", "in": "query", "schema": { "type": "string" } },
      "order": {
        "name": "order",
        "in": "query",
        "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }
      }
    },
    "schemas": {
      "Pedido": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "dataCompra": { "type": "string", "format": "date" }
        },
        "required": ["id", "dataCompra"]
      },
      "PedidoCreate": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "dataCompra": { "type": "string", "format": "date" }
        },
        "required": ["id", "dataCompra"]
      },
      "PedidoUpdate": {
        "type": "object",
        "properties": {
          "dataCompra": { "type": "string", "format": "date" }
        }
      },
      "PedidoPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Pedido" }
          },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" }
        }
      },
      "UploadJobCreated": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "data": { "type": "object" }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
      "PedidoStatus": {
        "type": "string",
        "enum": ["ABERTO", "FECHADO", "CANCELADO"]
      }
    }
  }
}